A mini game using arduino
It's a simple game I've made using Arduino, a 1.6" 5110 LCD and a joystick from 1:00 AM to 2:30 AM, so anyone can do it, faster and better. Feel free to do any change in codes, hardware, using them, publishing them in any way you like. You can also add some voice to it (it's not as easy as it sounds) and much more.

The game principle is so easy, in the beginning you have 10000 score, an small circle controlled by your joystick and another bigger circle that you have to catch by moving yours over it! pretty much that's it, also if you don't catch it in an specific time its position will change and you will get negative points, and by catching more points speed of game increases but if you lose points down to 10000 it will be reset again. the game ends in three ways, you loose all your points, the T.O that means you've ran out of time, or reaching 20500 that's nearly impossible!